Move AssetCopyService to the services package.
